Remove common dependency from library code
This removes the dependency from the dbhub.io/common module from the
library code by copying the required data types into the library. This
might look like a lot of code duplication, however it makes the two
projects more independent by not using data structures some of which are
solely for internal use and depending on implementation details on the
server side.

Another benefit is that users of the library do not have to import the
entire dbhub.io project including all of its dependencies into their
project.