Reduce dependencies on common code in test cases
This reduces the amount of dependencies on code in the dbhub.io/common
module in the test cases. The only remaining place where it is used is
for the diff function. Eventually this should also be reworked so we can
completely eliminate the import of the common module and have an
independent test here (i.e. not asking the server to do something, then
calling server code to verify its behaviour).